Move and remove old ts_admin sample code
